Changed type of GetArea() function in IShape from float to double

+++ Shapes/IShape.cs
@@ -15,6 +15,6 @@ namespace Shapes
         /// Calculate area of shape.
         /// </summary>
         /// <returns>Shape's area</returns>
-        public float GetArea();
+        public double GetArea();
     }
 }
